+++ packages/hub/src/account/typings/accountTypings.ts
@@ -1,0 +1,15 @@
+export type GetChainResultType = {
+  address: string;
+  chain: number;
+  chain_nodes: object;
+  ok: boolean;
+  result: 'found' | 'other_chain';
+  txtaddress: string;
+};
+
+export type LoginToWalletSagaInput = {
+  password?: string;
+  forceChain?: boolean;
+  address?: string;
+  wif?: string;
+}