+++ packages/hub/src/common/notification/typings/notificationTypings.ts
@@ -1,0 +1,8 @@
+import { AlertColor } from '@mui/material/Alert/Alert';
+
+export type NotificationType = {
+  text?: string;
+  type: AlertColor;
+  autoHideDuration?: number | null;
+  date?: string;
+};