+++ packages/tssdk/src/index.ts
@@ -1,1 +1,2 @@
 export { NetworkApi, AddressApi, instantiateSC, PaymentsApi, CryptoApi, TransactionsApi, SmartContractWrapper, loadScLocal } from './libs';
+export { ChainAction } from './helpers/network.enum';