+++ packages/tssdk/src/libs/network.ts
@@ -154,7 +154,7 @@ export class NetworkApi {
     return binaryCode;
   }
 
-  private async sendPreparedTX(tx: any, callback: Function, timeout: number) {
+  public async sendPreparedTX(tx: any, callback: Function, timeout: number = 1000, vm: 'wasm' | 'evm' = 'evm' ) {
     // await this.setChain(chain);
     const response = await this.askBlockchainTo(ChainAction.CREATE_TRANSACTION, { data: { tx } });
     if (callback) {
@@ -162,8 +162,6 @@ export class NetworkApi {
     }
     return response;
   }
-
-
 
   private calculateFeeSettings(settings: any) {
     let result = settings.current;
@@ -267,9 +265,8 @@ export class NetworkApi {
     }
   }
 
-  private async getAddressChain(address : string) {
-    const { chain } = await this.askBlockchainTo(ChainAction.GET_MY_CHAIN, { address });
-    return chain;
+  public async getAddressChain(address : string) {
+    return this.askBlockchainTo(ChainAction.GET_MY_CHAIN, { address });
   }
 
   private async askBlockchainTo(kind: ChainAction, parameters: any) {