+++ packages/tssdk/src/libs/transactions.ts
@@ -4,7 +4,6 @@ import * as msgPack from '@thepowereco/msgpack';
 const Bitcoin = require('bitcoinjs-lib');
 import { AddressApi } from './address';
 
-
 const TAG_PUBLIC_KEY = 0x02;
 const TAG_SIGNATURE = 0xFF;
 
@@ -294,8 +293,15 @@ export const TransactionsApi = {
     gasValue: number,
     wif: string,
     feeSettings: any,
+    vm: 'wasm' | 'evm' = 'wasm',
   ) {
     const selfInitParams = [Buffer.from(AddressApi.parseTextAddress(address))];
+    const scCode = vm === 'evm'
+      ? new Uint8Array(code.match(/[\da-f]{2}/gi).map(function (h: string) {
+        return parseInt(h, 16);
+      }))
+      : new Uint8Array(code);
+
     const body = {
       k: KIND_DEPLOY,
       t: +new Date(),
@@ -305,7 +311,7 @@ export const TransactionsApi = {
       p: [[PURPOSE_GAS, gasToken, gasValue]],
       c: ['init', selfInitParams],
       //"e": {'code': Buffer.from(new Uint8Array(code)), "vm": "wasm", "view": ["sha1:2b4ccea0d1de703012832f374e30effeff98fe4d", "/questions.wasm"]}
-      e: { 'code': Buffer.from(new Uint8Array(code)), 'vm': 'wasm', 'view': [] },
+      e: { 'code': Buffer.from(scCode), 'vm': vm, 'view': [] },
     };
     return TransactionsApi.packAndSignTX(computeFee(body, feeSettings), wif);
   },